BUGFIX Aroon TEST commit of AndroidWearSD; * ServiceStart according GoogleDocs and * https://stackoverflow.com/questions/55894636/android-9-pie-context-startforegroundservice-did-not-then-call-service-star/56338570#56338570 This commit is still in testing!
